Added a command-line option for loading a list of macros for which we should not expand. This option can be specified as follows:
mcpp -Z blocked_macros.txt test.Added a command-line option for loading
a list of macros for which we should not expand. This option can be
specified as follows:

mcpp -Z blocked_macros.txt test.c

I also increased the number of macros that can be defined via -D on the
command line.